+++ pkg/compactor/blocks_cleaner_test.go
@@ -19,6 +19,7 @@ import (
 	prom_testutil "github.com/prometheus/client_golang/prometheus/testutil"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
+	"github.com/thanos-io/thanos/pkg/block"
 	"github.com/thanos-io/thanos/pkg/block/metadata"
 	"github.com/thanos-io/thanos/pkg/objstore"
 
@@ -31,10 +32,11 @@ import (
 
 func TestBlocksCleaner(t *testing.T) {
 	for _, options := range []testBlocksCleanerOptions{
-		{concurrency: 1},
+		{concurrency: 1, tenantDeletionDelay: 0, user4FilesExist: false},
+		{concurrency: 1, tenantDeletionDelay: 2 * time.Hour, user4FilesExist: true},
+		{concurrency: 1, markersMigrationEnabled: true},
 		{concurrency: 2},
 		{concurrency: 10},
-		{concurrency: 1, markersMigrationEnabled: true},
 	} {
 		options := options
 
@@ -48,11 +50,13 @@ func TestBlocksCleaner(t *testing.T) {
 type testBlocksCleanerOptions struct {
 	concurrency             int
 	markersMigrationEnabled bool
+	tenantDeletionDelay     time.Duration
+	user4FilesExist         bool // User 4 has "FinishedTime" in tenant deletion marker set to "1h" ago.
 }
 
 func (o testBlocksCleanerOptions) String() string {
-	return fmt.Sprintf("concurrency=%d markers migration enabled=%v",
-		o.concurrency, o.markersMigrationEnabled)
+	return fmt.Sprintf("concurrency=%d, markers migration enabled=%v, tenant deletion delay=%v",
+		o.concurrency, o.markersMigrationEnabled, o.tenantDeletionDelay)
 }
 
 func testBlocksCleanerWithOptions(t *testing.T, options testBlocksCleanerOptions) {
@@ -85,9 +89,16 @@ func testBlocksCleanerWithOptions(t *testing.T, options testBlocksCleanerOptions
 	createDeletionMark(t, bucketClient, "user-2", block7, now.Add(-deletionDelay).Add(-time.Hour))            // Block reached the deletion threshold.
 
 	// Blocks for user-3, marked for deletion.
-	require.NoError(t, tsdb.WriteTenantDeletionMark(context.Background(), bucketClient, "user-3"))
+	require.NoError(t, tsdb.WriteTenantDeletionMark(context.Background(), bucketClient, "user-3", tsdb.NewTenantDeletionMark(time.Now())))
 	block9 := createTSDBBlock(t, bucketClient, "user-3", 10, 30, nil)
 	block10 := createTSDBBlock(t, bucketClient, "user-3", 30, 50, nil)
+
+	// User-4 with no more blocks, but couple of mark and debug files. Should be fully deleted.
+	user4Mark := tsdb.NewTenantDeletionMark(time.Now())
+	user4Mark.FinishedTime = time.Now().Unix() - 60 // Set to check final user cleanup.
+	require.NoError(t, tsdb.WriteTenantDeletionMark(context.Background(), bucketClient, "user-4", user4Mark))
+	user4DebugMetaFile := path.Join("user-4", block.DebugMetas, "meta.json")
+	require.NoError(t, bucketClient.Upload(context.Background(), user4DebugMetaFile, strings.NewReader("some random content here")))
 
 	// The fixtures have been created. If the bucket client wasn't wrapped to write
 	// deletion marks to the global location too, then this is the right time to do it.
@@ -100,6 +111,7 @@ func testBlocksCleanerWithOptions(t *testing.T, options testBlocksCleanerOptions
 		CleanupInterval:                    time.Minute,
 		CleanupConcurrency:                 options.concurrency,
 		BlockDeletionMarksMigrationEnabled: options.markersMigrationEnabled,
+		TenantCleanupDelay:                 options.tenantDeletionDelay,
 	}
 
 	reg := prometheus.NewPedanticRegistry()
@@ -138,6 +150,9 @@ func testBlocksCleanerWithOptions(t *testing.T, options testBlocksCleanerOptions
 		{path: path.Join("user-3", block10.String(), "index"), expectedExists: false},
 		// Tenant deletion mark is not removed.
 		{path: path.Join("user-3", tsdb.TenantDeletionMarkPath), expectedExists: true},
+		// User-4 is removed fully.
+		{path: path.Join("user-4", tsdb.TenantDeletionMarkPath), expectedExists: options.user4FilesExist},
+		{path: path.Join("user-4", block.DebugMetas, "meta.json"), expectedExists: options.user4FilesExist},
 	} {
 		exists, err := bucketClient.Exists(ctx, tc.path)
 		require.NoError(t, err)